additional test harness, partial fixes

+++ lib/src/main/java/tech/jnkr/presume/internal/atoms/Regular.java
@@ -1,3 +1,3 @@
 package tech.jnkr.presume.internal.atoms;
 
-public record Regular(float ratio, boolean sign, boolean simplify) implements DrawAtom {}
+public record Regular(double ratio, boolean sign, boolean simplify) implements DrawAtom {}
